+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Removed spurious warnings about vote content being not null.
+- Don't mark post as edited unless content and attachments are changed.
 
 ## [3.16.0] - 2025-02-03